:truck: move `DriverInfo` into its own module
